Full working PAD Controller Code, Freq only 2hz

+++ Firmware/ServoControllerFirmware/Core/Inc/thermocouple_utils.h
@@ -8,8 +8,8 @@
 #ifndef INC_THERMOCOUPLE_UTILS_H_
 #define INC_THERMOCOUPLE_UTILS_H_
 
-double Get_Temperature(uint32_t adc_val) {
-	double voltage = adc_val * (3.3 / 4096);
+float Get_Temperature(uint32_t adc_val) {
+	float voltage = adc_val * (3.3 / 4096);
 	return (voltage - 1.24) / 0.005;
 }
